refactor: separate SortByBottomSheet  into component
